[JENKINS-25111] Copy Artifacts from Phase Job
THIS IS AN INTEGRATION WITH Copy Artifact Plugin.
-Added a new Build Selector called MultiJobBuildSelector
to CopyArtifact Build Selector.
-If selected, this will make the Copy Artifact Plugin look
for an artifact from that job's build which was part of the
current MultiJobBuild.
-This selector will work properly if you configure it with
a MultiJobProject which has a PhaseJob that archives an
artifact that you want to retrieve as part of the MultiJob.
-This selector will also work if you configure it to Copy an
Artifact from another Phase Job which was part of the
current MultiJob Build.
-Adds CopyArtifact Plugin as an optional dependency.
-Adds Matrix Plugin as an optional dependency.