Apply the BuildFilter to the MultiJobBuildSelector
With this the MultiJobBuildSelector supports the additional BuildFilter
when matching builds. Using a BuildFilter currently is the only way to
choose between jobs of the same name that are triggered within the same
phase multiple times but with different parameters. The Copy Artifact
plugin can then use a parameter filter to choose between these jobs.