[FIXED JENKINS-30262] - Always use absolute URLs in MultiJob's JobColumn
The change replaces the implementation to Job's absolute url in order to always produce a valid link. The original approach does not work in general, because Views may not contain the multijob's items => relative paths may be broken.

Ridiculously, there's the same gloth in /lib/hudson/jobLink for the supported version core. It needs to be investigated separately