Fix problem when using folders
As mentioned in a comment of the JENKINS-25111 issue, the Multijob build selector does not work when using folders. The problem is due to the use of getFullDisplayName() instead of getDisplayName() when looking for a sub build. getFullDisplayName() includes the parent names whereas sub build's job name contains only the job name. It's the reason the check couldn't find a match.