allow sub-jobs to be queued using the standard jenkins build queue
this just adds a no-op build action to each sub-build object which makes
jenkins not consider them to be the same event, even if they are
triggering the same job. this allows us to not require waiting for each
sub-job to actually start before we can start the next one.