Support for NOT_BUILT build result
Previously a Multijob build would be flagged as ABORTED when a sub-build's result was
worse than FAILURE. A result of either NOT_BUILT or ABORTED would result in this. This
change adds a new isNotBuilt method so that the NOT_BUILT result is surfaced as the
result of a phase. The new logic results in ABORTED if the sub-build's result is
worse than NOT_BUILT and NOT_BUILT if the sub-build's result is worse than FAILURE.