Fix backwards compatibility bug with old build data
For heavy users of this plugin, upgrading the plugin exposes users to
an unhandled backwards incompatibility in the data persistence format.

In our case, it caused the OldDataMonitor object to become flooded with
error messages from trying to load old jobs, to the point of slowly but
surely filling up as much heap as we could throw at Jenkins due to the
large number of builds we have using this plugin.

This change implements the readResolve method used by the XStream
persistence layer to convert old build data to the new version, as
described in the Jenkins documentation. See the linked Jenkins help page
in code comments for more details.