Change  transient build field to private
The primary reason for the change is that when a MultiJob build goes in
the queue, Jenkins casts these objects to ParametersAction objects in
the shouldSchedule method in ParametersAction.java. These objects have a
deprecated private transient AbstractBuild build field and it seems that
setting this to a non-null, protected field during the cast gums up the
works further down the line, with the end result of the queue getting
completely stuck.

That's all in theory. We need to test it out to see if it fixes things
in practice.